Moved tower into scaffold.
