+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/world/Scaffold.kt
@@ -14,9 +14,11 @@ import net.ccbluex.liquidbounce.features.module.ModuleInfo
 import net.ccbluex.liquidbounce.features.module.modules.render.BlockOverlay
 import net.ccbluex.liquidbounce.ui.font.Fonts
 import net.ccbluex.liquidbounce.utils.*
+import net.ccbluex.liquidbounce.utils.block.BlockUtils
 import net.ccbluex.liquidbounce.utils.block.BlockUtils.canBeClicked
 import net.ccbluex.liquidbounce.utils.block.BlockUtils.isReplaceable
 import net.ccbluex.liquidbounce.utils.block.PlaceInfo
+import net.ccbluex.liquidbounce.utils.extensions.getBlock
 import net.ccbluex.liquidbounce.utils.render.RenderUtils
 import net.ccbluex.liquidbounce.utils.timer.MSTimer
 import net.ccbluex.liquidbounce.utils.timer.TickTimer
@@ -25,19 +27,19 @@ import net.ccbluex.liquidbounce.value.BoolValue
 import net.ccbluex.liquidbounce.value.FloatValue
 import net.ccbluex.liquidbounce.value.IntegerValue
 import net.ccbluex.liquidbounce.value.ListValue
-import net.minecraft.block.Block
 import net.minecraft.block.BlockBush
 import net.minecraft.client.gui.ScaledResolution
 import net.minecraft.client.renderer.GlStateManager
 import net.minecraft.client.renderer.RenderHelper
 import net.minecraft.client.settings.GameSettings
 import net.minecraft.init.Blocks
-import net.minecraft.init.Items
 import net.minecraft.item.ItemBlock
 import net.minecraft.item.ItemStack
+import net.minecraft.network.play.client.C03PacketPlayer
 import net.minecraft.network.play.client.C09PacketHeldItemChange
 import net.minecraft.network.play.client.C0APacketAnimation
 import net.minecraft.network.play.client.C0BPacketEntityAction
+import net.minecraft.stats.StatList
 import net.minecraft.util.BlockPos
 import net.minecraft.util.EnumFacing
 import net.minecraft.util.MathHelper.wrapAngleTo180_float
@@ -56,7 +58,7 @@ import kotlin.math.*
 )
 class Scaffold : Module() {
 
-    private val modeValue = ListValue("Mode", arrayOf("Normal", "Rewinside", "Expand"), "Normal")
+    private val scaffoldModeValue = ListValue("ScaffoldMode", arrayOf("Normal", "Rewinside", "Expand"), "Normal")
 
     // Delay
     private val maxDelayValue: IntegerValue = object : IntegerValue("MaxDelay", 0, 0, 1000) {
@@ -204,6 +206,41 @@ class Scaffold : Module() {
 
     private var itemStack = ItemStack(Blocks.barrier)
 
+    private val towerModeValue = ListValue(
+        "TowerMode",
+        arrayOf("Jump", "Motion", "ConstantMotion", "MotionTP", "Packet", "Teleport", "AAC3.3.9", "AAC3.6.4"),
+        "ConstantMotion"
+    )
+    private val stopWhenBlockAbove = BoolValue("StopWhenBlockAbove", false)
+    private val matrixValue = BoolValue("Matrix", false)
+
+    // Jump mode
+    private val jumpMotionValue = FloatValue("JumpMotion", 0.42f, 0.3681289f, 0.79f)
+    private val jumpDelayValue = IntegerValue("JumpDelay", 0, 0, 20)
+
+    // ConstantMotion
+    private val constantMotionValue = FloatValue("ConstantMotion", 0.42f, 0.1f, 1f)
+    private val constantMotionJumpGroundValue = FloatValue("ConstantMotionJumpGround", 0.79f, 0.76f, 1f)
+
+    // Teleport
+    private val teleportHeightValue = FloatValue("TeleportHeight", 1.15f, 0.1f, 5f)
+    private val teleportDelayValue = IntegerValue("TeleportDelay", 0, 0, 20)
+    private val teleportGroundValue = BoolValue("TeleportGround", true)
+    private val teleportNoMotionValue = BoolValue("TeleportNoMotion", false)
+
+    /**
+     * MODULE
+     */
+    // Target block
+    private var placeInfo: PlaceInfo? = null
+
+    // Mode stuff
+    private val timer = TickTimer()
+    private var jumpGround = 0.0
+
+    private var shouldTower = false
+
+
     // Enabling module
     override fun onEnable() {
         val player = mc.thePlayer ?: return
@@ -276,7 +313,7 @@ class Scaffold : Module() {
             }
         }
         if (player.onGround) {
-            when (modeValue.get().toLowerCase()) {
+            when (scaffoldModeValue.get().toLowerCase()) {
                 "rewinside" -> {
                     MovementUtils.strafe(0.2F)
                     player.motionY = 0.0
@@ -352,35 +389,174 @@ class Scaffold : Module() {
         val targetRotation = RotationUtils.targetRotation ?: return
         targetRotation.applyStrafeToPlayer(event)
         event.cancelEvent()
+
+
+    }
+
+    private fun fakeJump() {
+        mc.thePlayer!!.isAirBorne = true
+        mc.thePlayer!!.triggerAchievement(StatList.jumpStat)
     }
 
     @EventTarget
     fun onMotion(event: MotionEvent) {
-        val eventState = event.eventState
+        if (!shouldTower) {
+            val eventState = event.eventState
 
-        // Lock Rotation
-        if (rotationsValue.get() && (keepRotationValue.get() || !lockRotationTimer.hasTimePassed(keepLengthValue.get())) && lockRotation != null && strafeMode.get()
-                .equals("Off", true)
-        ) {
-            setRotation(lockRotation!!)
-            if (eventState == EventState.POST) {
-                lockRotationTimer.update()
+            // Lock Rotation
+            if (rotationsValue.get() && (keepRotationValue.get() || !lockRotationTimer.hasTimePassed(keepLengthValue.get())) && lockRotation != null && strafeMode.get()
+                    .equals("Off", true)
+            ) {
+                setRotation(lockRotation!!)
+                if (eventState == EventState.POST) {
+                    lockRotationTimer.update()
+                }
+            }
+
+            // Face block
+            if ((facesBlock || !rotationsValue.get()) && placeModeValue.get().equals(eventState.stateName, true)) {
+                place()
+            }
+
+            // Update and search for a new block
+            if (eventState == EventState.PRE && strafeMode.get().equals("Off", true)) {
+                update()
+            }
+
+            // Reset placeable delay
+            if (targetPlace == null && placeDelay.get()) {
+                delayTimer.reset()
+            }
+        } else {
+            val thePlayer = mc.thePlayer ?: return
+
+            // Lock Rotation
+            if (rotationsValue.get() && keepRotationValue.get() && lockRotation != null) {
+                RotationUtils.setTargetRotation(lockRotation)
+            }
+
+            mc.timer.timerSpeed = timerValue.get()
+            val eventState = event.eventState
+
+            if (placeModeValue.get().equals(eventState.stateName, ignoreCase = true)) {
+                place()
+            }
+
+            if (eventState == EventState.PRE) {
+                placeInfo = null
+                timer.update()
+
+                val update = if (!autoBlockValue.get().equals("Off", ignoreCase = true)) {
+                    InventoryUtils.findAutoBlockBlock(randomSlotValue.get()) != -1 || thePlayer.heldItem != null && thePlayer.heldItem!!.item is ItemBlock
+                } else {
+                    thePlayer.heldItem != null && thePlayer.heldItem!!.item is ItemBlock
+                }
+
+                if (update) {
+                    if (!stopWhenBlockAbove.get() || BlockUtils.getBlock(
+                            BlockPos(
+                                thePlayer.posX, thePlayer.posY + 2, thePlayer.posZ
+                            )
+                        ) == Blocks.air
+                    ) {
+                        move()
+                    }
+                    val blockPos = BlockPos(thePlayer.posX, thePlayer.posY - 1.0, thePlayer.posZ)
+                    if (blockPos.getBlock() == Blocks.air) {
+                        if (search(blockPos,false) && rotationsValue.get()) {
+                            val vecRotation = RotationUtils.faceBlock(blockPos)
+                            if (vecRotation != null) {
+                                RotationUtils.setTargetRotation(vecRotation.rotation)
+                                placeInfo!!.vec3 = vecRotation.vec
+                            }
+                        }
+                    }
+                }
             }
         }
+    }
 
-        // Face block
-        if ((facesBlock || !rotationsValue.get()) && placeModeValue.get().equals(eventState.stateName, true)) {
-            place()
-        }
+    private fun move() {
+        val thePlayer = mc.thePlayer ?: return
 
-        // Update and search for a new block
-        if (eventState == EventState.PRE && strafeMode.get().equals("Off", true)) {
-            update()
-        }
-
-        // Reset placeable delay
-        if (targetPlace == null && placeDelay.get()) {
-            delayTimer.reset()
+        when (towerModeValue.get().toLowerCase()) {
+            "jump" -> if (thePlayer.onGround && timer.hasTimePassed(jumpDelayValue.get())) {
+                fakeJump()
+                thePlayer.motionY = jumpMotionValue.get().toDouble()
+                timer.reset()
+            }
+            "motion" -> if (thePlayer.onGround) {
+                fakeJump()
+                thePlayer.motionY = 0.42
+            } else if (thePlayer.motionY < 0.1) {
+                thePlayer.motionY = -0.3
+            }
+            "motiontp" -> if (thePlayer.onGround) {
+                fakeJump()
+                thePlayer.motionY = 0.42
+            } else if (thePlayer.motionY < 0.23) {
+                thePlayer.setPosition(thePlayer.posX, truncate(thePlayer.posY), thePlayer.posZ)
+            }
+            "packet" -> if (thePlayer.onGround && timer.hasTimePassed(2)) {
+                fakeJump()
+                mc.netHandler.addToSendQueue(
+                    C03PacketPlayer.C04PacketPlayerPosition(
+                        thePlayer.posX, thePlayer.posY + 0.42, thePlayer.posZ, false
+                    )
+                )
+                mc.netHandler.addToSendQueue(
+                    C03PacketPlayer.C04PacketPlayerPosition(
+                        thePlayer.posX, thePlayer.posY + 0.753, thePlayer.posZ, false
+                    )
+                )
+                thePlayer.setPosition(thePlayer.posX, thePlayer.posY + 1.0, thePlayer.posZ)
+                timer.reset()
+            }
+            "teleport" -> {
+                if (teleportNoMotionValue.get()) {
+                    thePlayer.motionY = 0.0
+                }
+                if ((thePlayer.onGround || !teleportGroundValue.get()) && timer.hasTimePassed(teleportDelayValue.get())) {
+                    fakeJump()
+                    thePlayer.setPositionAndUpdate(
+                        thePlayer.posX, thePlayer.posY + teleportHeightValue.get(), thePlayer.posZ
+                    )
+                    timer.reset()
+                }
+            }
+            "constantmotion" -> {
+                if (thePlayer.onGround) {
+                    fakeJump()
+                    jumpGround = thePlayer.posY
+                    thePlayer.motionY = constantMotionValue.get().toDouble()
+                }
+                if (thePlayer.posY > jumpGround + constantMotionJumpGroundValue.get()) {
+                    fakeJump()
+                    thePlayer.setPosition(
+                        thePlayer.posX, truncate(thePlayer.posY), thePlayer.posZ
+                    ) // TODO: toInt() required?
+                    thePlayer.motionY = constantMotionValue.get().toDouble()
+                    jumpGround = thePlayer.posY
+                }
+            }
+            "aac3.3.9" -> {
+                if (thePlayer.onGround) {
+                    fakeJump()
+                    thePlayer.motionY = 0.4001
+                }
+                mc.timer.timerSpeed = 1f
+                if (thePlayer.motionY < 0) {
+                    thePlayer.motionY -= 0.00000945
+                    mc.timer.timerSpeed = 1.6f
+                }
+            }
+            "aac3.6.4" -> if (thePlayer.ticksExisted % 4 == 1) {
+                thePlayer.motionY = 0.4195464
+                thePlayer.setPosition(thePlayer.posX - 0.035, thePlayer.posY, thePlayer.posZ)
+            } else if (thePlayer.ticksExisted % 4 == 0) {
+                thePlayer.motionY = -0.5
+                thePlayer.setPosition(thePlayer.posX + 0.035, thePlayer.posY, thePlayer.posZ)
+            }
         }
     }
 
@@ -395,7 +571,7 @@ class Scaffold : Module() {
             return
         }
 
-        findBlock(modeValue.get().equals("expand", true))
+        findBlock(scaffoldModeValue.get().equals("expand", true))
     }
 
     private fun setRotation(rotation: Rotation) {
@@ -451,75 +627,132 @@ class Scaffold : Module() {
     }
 
     fun place() {
-        val player = mc.thePlayer ?: return
-        val world = mc.theWorld ?: return
+        if (shouldTower) {
+            if (placeInfo == null) return
+            val thePlayer = mc.thePlayer ?: return
 
-        if (targetPlace == null) {
-            if (placeDelay.get()) {
-                delayTimer.reset()
+            // AutoBlock
+            if (thePlayer.heldItem != null) {
+                itemStack = thePlayer.heldItem
             }
-            return
-        }
 
-        if (!delayTimer.hasTimePassed(delay) || sameYValue.get() && launchY - 1 != targetPlace!!.vec3.yCoord.toInt()) {
-            return
-        }
+            if (thePlayer.heldItem == null || itemStack.item !is ItemBlock || (itemStack.item as ItemBlock).block is BlockBush) {
+                val blockSlot = InventoryUtils.findAutoBlockBlock(randomSlotValue.get())
 
-        itemStack = player.heldItem
-        if (itemStack == null || itemStack.item !is ItemBlock|| (itemStack.item!! as ItemBlock).block is BlockBush || player.heldItem!!.stackSize <= 0) {
-            blockSlot = InventoryUtils.findAutoBlockBlock(randomSlotValue.get())
+                if (blockSlot == -1) return
 
-            if (blockSlot == -1) {
+                when (autoBlockValue.get()) {
+                    "Off" -> return
+                    "Pick" -> {
+                        mc.thePlayer!!.inventory.currentItem = blockSlot - 36
+                        mc.playerController.updateController()
+                    }
+                    "Spoof" -> {
+                        if (blockSlot - 36 != slot) {
+                            mc.netHandler.addToSendQueue(C09PacketHeldItemChange(blockSlot - 36))
+                        }
+                    }
+                    "Switch" -> {
+                        if (blockSlot - 36 != slot) {
+                            mc.netHandler.addToSendQueue(C09PacketHeldItemChange(blockSlot - 36))
+                        }
+                    }
+                }
+                itemStack = thePlayer.inventoryContainer.getSlot(blockSlot).stack
+            }
+
+            // Place block
+            if (mc.playerController.onPlayerRightClick(
+                    thePlayer, mc.theWorld!!, itemStack!!, placeInfo!!.blockPos, placeInfo!!.enumFacing, placeInfo!!.vec3
+                )
+            ) {
+                if (swingValue.get()) {
+                    thePlayer.swingItem()
+                } else {
+                    mc.netHandler.addToSendQueue(C0APacketAnimation())
+                }
+            }
+            if (autoBlockValue.get().equals("Switch", true)) {
+                if (slot != mc.thePlayer!!.inventory.currentItem) mc.netHandler.addToSendQueue(
+                    C09PacketHeldItemChange(
+                        mc.thePlayer!!.inventory.currentItem
+                    )
+                )
+            }
+            placeInfo = null
+        } else {
+            val player = mc.thePlayer ?: return
+            val world = mc.theWorld ?: return
+
+            if (targetPlace == null) {
+                if (placeDelay.get()) {
+                    delayTimer.reset()
+                }
                 return
             }
 
-            when (autoBlockValue.get().toLowerCase()) {
-                "off" -> {
+            if (!delayTimer.hasTimePassed(delay) || sameYValue.get() && launchY - 1 != targetPlace!!.vec3.yCoord.toInt()) {
+                return
+            }
+
+            if (mc.thePlayer.heldItem != null) {
+                itemStack = mc.thePlayer.heldItem
+            }
+            if (player.heldItem == null || itemStack.item !is ItemBlock|| (itemStack.item!! as ItemBlock).block is BlockBush || player.heldItem!!.stackSize <= 0) {
+                blockSlot = InventoryUtils.findAutoBlockBlock(randomSlotValue.get())
+
+                if (blockSlot == -1) {
                     return
                 }
-                "pick" -> {
-                    player.inventory.currentItem = blockSlot - 36
-                    mc.playerController.updateController()
-                }
-                "spoof" -> {
-                    if (blockSlot - 36 != slot) {
-                        mc.netHandler.addToSendQueue(C09PacketHeldItemChange(blockSlot - 36))
+
+                when (autoBlockValue.get().toLowerCase()) {
+                    "off" -> {
+                        return
+                    }
+                    "pick" -> {
+                        player.inventory.currentItem = blockSlot - 36
+                        mc.playerController.updateController()
+                    }
+                    "spoof" -> {
+                        if (blockSlot - 36 != slot) {
+                            mc.netHandler.addToSendQueue(C09PacketHeldItemChange(blockSlot - 36))
+                        }
+                    }
+                    "switch" -> {
+                        if (blockSlot - 36 != slot) {
+                            mc.netHandler.addToSendQueue(C09PacketHeldItemChange(blockSlot - 36))
+                        }
                     }
                 }
-                "switch" -> {
-                    if (blockSlot - 36 != slot) {
-                        mc.netHandler.addToSendQueue(C09PacketHeldItemChange(blockSlot - 36))
-                    }
+                itemStack = player.inventoryContainer.getSlot(blockSlot).stack
+            }
+
+            if (mc.playerController.onPlayerRightClick(
+                    player, world, itemStack, targetPlace!!.blockPos, targetPlace!!.enumFacing, targetPlace!!.vec3
+                )
+            ) {
+                delayTimer.reset()
+                delay = if (!placeDelay.get()) 0 else TimeUtils.randomDelay(minDelayValue.get(), maxDelayValue.get())
+
+                if (player.onGround) {
+                    val modifier = speedModifierValue.get()
+                    player.motionX = player.motionX * modifier
+                    player.motionZ = player.motionZ * modifier
+                }
+
+                if (swingValue.get()) {
+                    player.swingItem()
+                } else {
+                    mc.netHandler.addToSendQueue(C0APacketAnimation())
                 }
             }
-            itemStack = player.inventoryContainer.getSlot(blockSlot).stack
-        }
-
-        if (mc.playerController.onPlayerRightClick(
-                player, world, itemStack, targetPlace!!.blockPos, targetPlace!!.enumFacing, targetPlace!!.vec3
-            )
-        ) {
-            delayTimer.reset()
-            delay = if (!placeDelay.get()) 0 else TimeUtils.randomDelay(minDelayValue.get(), maxDelayValue.get())
-
-            if (player.onGround) {
-                val modifier = speedModifierValue.get()
-                player.motionX = player.motionX * modifier
-                player.motionZ = player.motionZ * modifier
+            if (autoBlockValue.get().equals("Switch", true)) {
+                if (slot != player.inventory.currentItem) {
+                    mc.netHandler.addToSendQueue(C09PacketHeldItemChange(player.inventory.currentItem))
+                }
             }
-
-            if (swingValue.get()) {
-                player.swingItem()
-            } else {
-                mc.netHandler.addToSendQueue(C0APacketAnimation())
-            }
+            targetPlace = null
         }
-        if (autoBlockValue.get().equals("Switch", true)) {
-            if (slot != player.inventory.currentItem) {
-                mc.netHandler.addToSendQueue(C09PacketHeldItemChange(player.inventory.currentItem))
-            }
-        }
-        targetPlace = null
     }
 
     // Disabling module
@@ -604,6 +837,11 @@ class Scaffold : Module() {
 
             GL11.glPopMatrix()
         }
+        if (mc.gameSettings.keyBindJump.isKeyDown) {
+            shouldTower = true
+        } else {
+            shouldTower = false
+        }
     }
 
     // Visuals
@@ -613,7 +851,7 @@ class Scaffold : Module() {
         if (!markValue.get()) {
             return
         }
-        for (i in 0 until if (modeValue.get().equals("Expand", true)) expandLengthValue.get() + 1 else 2) {
+        for (i in 0 until if (scaffoldModeValue.get().equals("Expand", true)) expandLengthValue.get() + 1 else 2) {
             val yaw = Math.toRadians(player.rotationYaw.toDouble())
             val x = if (omniDirectionalExpand.get()) -sin(yaw).roundToInt() else player.horizontalFacing.directionVec.x
             val z = if (omniDirectionalExpand.get()) cos(yaw).roundToInt() else player.horizontalFacing.directionVec.z
@@ -628,6 +866,7 @@ class Scaffold : Module() {
                 break
             }
         }
+
     }
 
     /**
@@ -639,135 +878,218 @@ class Scaffold : Module() {
      */
 
     private fun search(blockPosition: BlockPos, raycast: Boolean): Boolean {
-        facesBlock = false
-        val player = mc.thePlayer ?: return false
-        val world = mc.theWorld ?: return false
-
-        if (!isReplaceable(blockPosition)) {
-            return false
-        }
-
-        // Search Ranges
-        val xzRV = xzRangeValue.get().toDouble()
-        val xzSSV = calcStepSize(xzRV.toFloat())
-        val yRV = yRangeValue.get().toDouble()
-        val ySSV = calcStepSize(yRV.toFloat())
-        val eyesPos = Vec3(player.posX, player.entityBoundingBox.minY + player.eyeHeight, player.posZ)
-        var placeRotation: PlaceRotation? = null
-        for (facingType in EnumFacing.values()) {
-            val neighbor = blockPosition.offset(facingType)
-            if (!canBeClicked(neighbor)) {
-                continue
+        if (shouldTower) {
+            val thePlayer = mc.thePlayer ?: return false
+            if (!isReplaceable(blockPosition)) {
+                return false
             }
-            val dirVec = Vec3(facingType.directionVec)
-            val auto = searchMode.get().equals("Auto", true)
-            val center = searchMode.get().equals("AutoCenter", true)
-            var xSearch = if (auto) 0.1 else 0.5 - xzRV / 2
-            while (xSearch <= if (auto) 0.9 else 0.5 + xzRV / 2) {
-                var ySearch = if (auto) 0.1 else 0.5 - yRV / 2
-                while (ySearch <= if (auto) 0.9 else 0.5 + yRV / 2) {
-                    var zSearch = if (auto) 0.1 else 0.5 - xzRV / 2
-                    while (zSearch <= if (auto) 0.9 else 0.5 + xzRV / 2) {
-                        val posVec = Vec3(blockPosition).addVector(
-                            if (center) 0.5 else xSearch, if (center) 0.5 else ySearch, if (center) 0.5 else zSearch
-                        )
-                        val distanceSqPosVec = eyesPos.squareDistanceTo(posVec)
-                        val hitVec = posVec.add(Vec3(dirVec.xCoord * 0.5, dirVec.yCoord * 0.5, dirVec.zCoord * 0.5))
-                        if (raycast && (eyesPos.distanceTo(hitVec) > 4.25 || distanceSqPosVec > eyesPos.squareDistanceTo(
-                                posVec.add(dirVec)
-                            ) || world.rayTraceBlocks(
-                                eyesPos,
-                                hitVec,
-                                false,
-                                true,
-                                false
-                            ) != null)
-                        ) {
-                            zSearch += if (auto) 0.1 else xzSSV
-                            continue
-                        }
 
-                        // Face block
-                        val diffX = hitVec.xCoord - eyesPos.xCoord
-                        val diffY = hitVec.yCoord - eyesPos.yCoord
-                        val diffZ = hitVec.zCoord - eyesPos.zCoord
-                        val diffXZ = sqrt(diffX * diffX + diffZ * diffZ)
-                        if (facingType != EnumFacing.UP && facingType != EnumFacing.DOWN) {
-                            val diff = abs(if (facingType == EnumFacing.NORTH || facingType == EnumFacing.SOUTH) diffZ else diffX)
-                            if (diff < minDistValue.get()) {
+            val eyesPos = Vec3(thePlayer.posX, thePlayer.entityBoundingBox.minY + thePlayer.eyeHeight, thePlayer.posZ)
+            var placeRotation: PlaceRotation? = null
+            for (facingType in EnumFacing.values()) {
+                val neighbor = blockPosition.offset(facingType)
+                if (!canBeClicked(neighbor)) {
+                    continue
+                }
+                val dirVec = Vec3(facingType.directionVec)
+                val matrix = matrixValue.get()
+                var xSearch = 0.1
+                while (xSearch < 0.9) {
+                    var ySearch = 0.1
+                    while (ySearch < 0.9) {
+                        var zSearch = 0.1
+                        while (zSearch < 0.9) {
+                            val posVec = Vec3(blockPosition).addVector(
+                                if (matrix) 0.5 else xSearch, if (matrix) 0.5 else ySearch, if (matrix) 0.5 else zSearch
+                            )
+                            val distanceSqPosVec = eyesPos.squareDistanceTo(posVec)
+                            val hitVec = posVec.add(Vec3(dirVec.xCoord * 0.5, dirVec.yCoord * 0.5, dirVec.zCoord * 0.5))
+                            if (eyesPos.distanceTo(hitVec) > 4.25 || distanceSqPosVec > eyesPos.squareDistanceTo(
+                                    posVec.add(dirVec)
+                                ) || mc.theWorld!!.rayTraceBlocks(
+                                    eyesPos,
+                                    hitVec,
+                                    false,
+                                    true,
+                                    false
+                                ) != null
+                            ) {
+                                zSearch += 0.1
+                                continue
+                            }
+
+                            // face block
+                            val rotation = RotationUtils.toRotation(hitVec, false)
+
+                            val rotationVector = RotationUtils.getVectorForRotation(rotation)
+                            val vector = eyesPos.addVector(
+                                rotationVector.xCoord * 4.25, rotationVector.yCoord * 4.25, rotationVector.zCoord * 4.25
+                            )
+                            val obj = mc.theWorld!!.rayTraceBlocks(
+                                eyesPos,
+                                vector,
+                                false,
+                                false,
+                                true
+                            ) ?: continue
+
+                            if (obj.typeOfHit != MovingObjectPosition.MovingObjectType.BLOCK || obj.blockPos != neighbor) {
+                                zSearch += 0.1
+                                continue
+                            }
+
+                            if (placeRotation == null || RotationUtils.getRotationDifference(rotation) < RotationUtils.getRotationDifference(
+                                    placeRotation.rotation
+                                )
+                            ) {
+                                placeRotation = PlaceRotation(PlaceInfo(neighbor, facingType.opposite, hitVec), rotation)
+                            }
+                            zSearch += 0.1
+                        }
+                        ySearch += 0.1
+                    }
+                    xSearch += 0.1
+                }
+            }
+            if (placeRotation == null) return false
+            if (rotationsValue.get()) {
+                RotationUtils.setTargetRotation(placeRotation.rotation, 0)
+                lockRotation = placeRotation.rotation
+            }
+            placeInfo = placeRotation.placeInfo
+            return true
+        } else {
+            facesBlock = false
+            val player = mc.thePlayer ?: return false
+            val world = mc.theWorld ?: return false
+
+            if (!isReplaceable(blockPosition)) {
+                return false
+            }
+
+            // Search Ranges
+            val xzRV = xzRangeValue.get().toDouble()
+            val xzSSV = calcStepSize(xzRV.toFloat())
+            val yRV = yRangeValue.get().toDouble()
+            val ySSV = calcStepSize(yRV.toFloat())
+            val eyesPos = Vec3(player.posX, player.entityBoundingBox.minY + player.eyeHeight, player.posZ)
+            var placeRotation: PlaceRotation? = null
+            for (facingType in EnumFacing.values()) {
+                val neighbor = blockPosition.offset(facingType)
+                if (!canBeClicked(neighbor)) {
+                    continue
+                }
+                val dirVec = Vec3(facingType.directionVec)
+                val auto = searchMode.get().equals("Auto", true)
+                val center = searchMode.get().equals("AutoCenter", true)
+                var xSearch = if (auto) 0.1 else 0.5 - xzRV / 2
+                while (xSearch <= if (auto) 0.9 else 0.5 + xzRV / 2) {
+                    var ySearch = if (auto) 0.1 else 0.5 - yRV / 2
+                    while (ySearch <= if (auto) 0.9 else 0.5 + yRV / 2) {
+                        var zSearch = if (auto) 0.1 else 0.5 - xzRV / 2
+                        while (zSearch <= if (auto) 0.9 else 0.5 + xzRV / 2) {
+                            val posVec = Vec3(blockPosition).addVector(
+                                if (center) 0.5 else xSearch, if (center) 0.5 else ySearch, if (center) 0.5 else zSearch
+                            )
+                            val distanceSqPosVec = eyesPos.squareDistanceTo(posVec)
+                            val hitVec = posVec.add(Vec3(dirVec.xCoord * 0.5, dirVec.yCoord * 0.5, dirVec.zCoord * 0.5))
+                            if (raycast && (eyesPos.distanceTo(hitVec) > 4.25 || distanceSqPosVec > eyesPos.squareDistanceTo(
+                                    posVec.add(dirVec)
+                                ) || world.rayTraceBlocks(
+                                    eyesPos,
+                                    hitVec,
+                                    false,
+                                    true,
+                                    false
+                                ) != null)
+                            ) {
                                 zSearch += if (auto) 0.1 else xzSSV
                                 continue
                             }
-                        }
-                        val rotation = Rotation(
-                            wrapAngleTo180_float(Math.toDegrees(atan2(diffZ, diffX)).toFloat() - 90f),
-                            wrapAngleTo180_float(-Math.toDegrees(atan2(diffY, diffXZ)).toFloat())
-                        )
-                        val rotationVector = RotationUtils.getVectorForRotation(rotation)
-                        val vector = eyesPos.addVector(
-                            rotationVector.xCoord * 4.25, rotationVector.yCoord * 4.25, rotationVector.zCoord * 4.25
-                        )
 
-                        val obj = world.rayTraceBlocks(
-                            eyesPos,
-                            vector,
-                            false,
-                            false,
-                            true
-                        ) ?: continue
-
-                        if (obj.typeOfHit != MovingObjectPosition.MovingObjectType.BLOCK || obj.blockPos != neighbor) {
-                            zSearch += if (auto) 0.1 else xzSSV
-                            continue
-                        }
-                        if (placeRotation == null || RotationUtils.getRotationDifference(rotation) < RotationUtils.getRotationDifference(
-                                placeRotation.rotation
+                            // Face block
+                            val diffX = hitVec.xCoord - eyesPos.xCoord
+                            val diffY = hitVec.yCoord - eyesPos.yCoord
+                            val diffZ = hitVec.zCoord - eyesPos.zCoord
+                            val diffXZ = sqrt(diffX * diffX + diffZ * diffZ)
+                            if (facingType != EnumFacing.UP && facingType != EnumFacing.DOWN) {
+                                val diff = abs(if (facingType == EnumFacing.NORTH || facingType == EnumFacing.SOUTH) diffZ else diffX)
+                                if (diff < minDistValue.get()) {
+                                    zSearch += if (auto) 0.1 else xzSSV
+                                    continue
+                                }
+                            }
+                            val rotation = Rotation(
+                                wrapAngleTo180_float(Math.toDegrees(atan2(diffZ, diffX)).toFloat() - 90f),
+                                wrapAngleTo180_float(-Math.toDegrees(atan2(diffY, diffXZ)).toFloat())
                             )
-                        ) {
-                            placeRotation = PlaceRotation(PlaceInfo(neighbor, facingType.opposite, hitVec), rotation)
+                            val rotationVector = RotationUtils.getVectorForRotation(rotation)
+                            val vector = eyesPos.addVector(
+                                rotationVector.xCoord * 4.25, rotationVector.yCoord * 4.25, rotationVector.zCoord * 4.25
+                            )
+
+                            val obj = world.rayTraceBlocks(
+                                eyesPos,
+                                vector,
+                                false,
+                                false,
+                                true
+                            ) ?: continue
+
+                            if (obj.typeOfHit != MovingObjectPosition.MovingObjectType.BLOCK || obj.blockPos != neighbor) {
+                                zSearch += if (auto) 0.1 else xzSSV
+                                continue
+                            }
+                            if (placeRotation == null || RotationUtils.getRotationDifference(rotation) < RotationUtils.getRotationDifference(
+                                    placeRotation.rotation
+                                )
+                            ) {
+                                placeRotation = PlaceRotation(PlaceInfo(neighbor, facingType.opposite, hitVec), rotation)
+                            }
+
+                            zSearch += if (auto) 0.1 else xzSSV
                         }
-
-                        zSearch += if (auto) 0.1 else xzSSV
+                        ySearch += if (auto) 0.1 else ySSV
                     }
-                    ySearch += if (auto) 0.1 else ySSV
+                    xSearch += if (auto) 0.1 else xzSSV
                 }
-                xSearch += if (auto) 0.1 else xzSSV
             }
-        }
-        if (placeRotation == null) {
-            return false
-        }
-        if (rotationsValue.get()) {
-            if (minTurnSpeedValue.get() < 180) {
-                val limitedRotation = RotationUtils.limitAngleChange(
-                    RotationUtils.serverRotation,
-                    placeRotation.rotation,
-                    (Math.random() * (maxTurnSpeedValue.get() - minTurnSpeedValue.get()) + minTurnSpeedValue.get()).toFloat()
-                )
+            if (placeRotation == null) {
+                return false
+            }
+            if (rotationsValue.get()) {
+                if (minTurnSpeedValue.get() < 180) {
+                    val limitedRotation = RotationUtils.limitAngleChange(
+                        RotationUtils.serverRotation,
+                        placeRotation.rotation,
+                        (Math.random() * (maxTurnSpeedValue.get() - minTurnSpeedValue.get()) + minTurnSpeedValue.get()).toFloat()
+                    )
 
-                if ((10 * wrapAngleTo180_float(limitedRotation.yaw)).roundToInt() == (10 * wrapAngleTo180_float(
-                        placeRotation.rotation.yaw
-                    )).roundToInt() && (10 * wrapAngleTo180_float(limitedRotation.pitch)).roundToInt() == (10 * wrapAngleTo180_float(
-                        placeRotation.rotation.pitch
-                    )).roundToInt()
-                ) {
+                    if ((10 * wrapAngleTo180_float(limitedRotation.yaw)).roundToInt() == (10 * wrapAngleTo180_float(
+                            placeRotation.rotation.yaw
+                        )).roundToInt() && (10 * wrapAngleTo180_float(limitedRotation.pitch)).roundToInt() == (10 * wrapAngleTo180_float(
+                            placeRotation.rotation.pitch
+                        )).roundToInt()
+                    ) {
+                        setRotation(placeRotation.rotation)
+                        lockRotation = placeRotation.rotation
+                        facesBlock = true
+                    } else {
+                        setRotation(limitedRotation)
+                        lockRotation = limitedRotation
+                        facesBlock = false
+                    }
+                } else {
                     setRotation(placeRotation.rotation)
                     lockRotation = placeRotation.rotation
                     facesBlock = true
-                } else {
-                    setRotation(limitedRotation)
-                    lockRotation = limitedRotation
-                    facesBlock = false
                 }
-            } else {
-                setRotation(placeRotation.rotation)
-                lockRotation = placeRotation.rotation
-                facesBlock = true
+                lockRotationTimer.reset()
             }
-            lockRotationTimer.reset()
+            targetPlace = placeRotation.placeInfo
+            return true
         }
-        targetPlace = placeRotation.placeInfo
-        return true
+
     }
 
     private fun calcStepSize(range: Float): Double {
@@ -794,5 +1116,5 @@ class Scaffold : Module() {
             return amount
         }
     override val tag: String
-        get() = modeValue.get()
+        get() = scaffoldModeValue.get()
 }
